refactor(migrations): unify _migrations.py into _migrations/ package
The previous commit added src/gradata/_migrations/ as a package but the
sibling src/gradata/_migrations.py still shadowed it, so
Brain.__init__ crashed with "cannot import name 'run_migrations'".

This commit merges both:

- _migrations/__init__.py exposes run_migrations(db_path) with the
  signature Brain.__init__ expects.
- Keeps the inline _BASE_TABLES + _MIGRATIONS idempotent ALTER list for
  the fast path on existing brains.
- Adds tracked numbered-migration dispatch: scans NNN_*.py in the
  package, skips any already in the migrations table, and invokes
  module.up(conn, tenant_id). Tenant UUID resolved via
  get_or_create_tenant_id on the brain dir.
- Deletes the obsolete _migrations.py.

Brain.init now runs both layers on every open, so upgrades are
transparent.
